[TPU] Make triton an optional import on the core run path
Non-Triton devices (cpu, tpu) can import and run tritonbench without triton
installed. triton is only used inside Triton/CUDA timing paths, so:

- guard `import triton` with try/except in env_utils, triton_op, do_bench/run
- lazy-import the CUDA-only timing backends (gpu_events, power) inside their
  do_bench_wrapper branches
- move the AMD `sleep_amd` @triton.jit kernel from gpu_utils (broadly imported
  on the run path) into gpu_events, its only consumer (CUDA-only)

Also fixes `--device cpu` requiring triton.
